src/main/java/ca/thoughtwire/readyapi/testresulttransfer/services/TestCaseService.java
package ca.thoughtwire.readyapi.testresulttransfer.services;

import ca.thoughtwire.readyapi.testresulttransfer.models.TestCase;
import ca.thoughtwire.readyapi.testresulttransfer.repositories.TestCaseRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class TestCaseService {

    @Autowired
    private TestCaseRepository testCaseRepository;

    public List<TestCase> list() {
        return testCaseRepository.findAll();
    }

}
